Add a very basic performance test
This consists of a server and a client application that will performing
testing using benchmark.js. To run it, execute the following command

  $ node test/perf/server.js

and navigate to http://localhost:8080 in your favourite browser. The
implementation is currently compared against the native implementation
and Google's wgxpath.

The preliminary results show that my implementation is somewhat slower
than Google's.

  document#evaluate x 250,487 ops/sec ±0.93% (65 runs sampled)
  XPathDOM#evaluate x 24,319 ops/sec ±3.34% (67 runs sampled)
  wgxpath#evaluate x 77,202 ops/sec ±1.97% (61 runs sampled)
  Fastest is document#evaluate