Replace shrinkwrap with a package-lock
This locks down dev dependencies and will ensure that `npm install` on
Travis, with a hot cache, takes significantly less time than it used to.
It also unforunately resolves all dependencies again, which kind of
defeats the purpose of having a lock file.

Moreover, it changes the builds slightly, due to the above mentioned
update in dependencies (EG. `xpath-analyzer` used to be locked to 1.0.0,
but is now at 1.1.2).